+++ lib/ui/news/screen.dart
@@ -1,62 +1,81 @@
-import 'dart:async';
-
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:http/http.dart' as http;
-import 'package:native_widgets/native_widgets.dart';
-import 'package:webfeed/webfeed.dart';
 
-class NewsScreen extends StatelessWidget {
-  final GlobalKey<ScaffoldState> _scaffoldKey = GlobalKey<ScaffoldState>();
+import '../../bloc/bloc.dart';
+import '../app/bottom_loader.dart';
+import 'item.dart';
+
+class ArticlesScreen extends StatefulWidget {
+  @override
+  _ArticlesScreenState createState() => _ArticlesScreenState();
+}
+
+class _ArticlesScreenState extends State<ArticlesScreen> {
+  final _scrollController = ScrollController();
+  final PostBloc _postBloc = PostBloc(httpClient: http.Client());
+  final _scrollThreshold = 200.0;
+
+  _ArticlesScreenState() {
+    _scrollController.addListener(_onScroll);
+    _postBloc.dispatch(Fetch());
+  }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      key: _scaffoldKey,
       appBar: AppBar(
-        title: const Text('Breaking News'),
-        centerTitle: true,
+        title: Text("Space News"),
       ),
-      body: NewsList(),
+      body: BlocBuilder(
+        bloc: _postBloc,
+        builder: (BuildContext context, PostState state) {
+          if (state is PostUninitialized) {
+            return Center(
+              child: CircularProgressIndicator(),
+            );
+          }
+          if (state is PostError) {
+            return Center(
+              child: Text('failed to fetch posts'),
+            );
+          }
+          if (state is PostLoaded) {
+            if (state.posts.isEmpty) {
+              return Center(
+                child: Text('no posts'),
+              );
+            }
+            return ListView.builder(
+              itemBuilder: (BuildContext context, int index) {
+                if (index >= state.posts.length) {
+                  return BottomLoader();
+                } else {
+                  return PostWidget(post: state.posts[index]);
+                }
+              },
+              itemCount: state.hasReachedMax
+                  ? state.posts.length
+                  : state.posts.length + 1,
+              controller: _scrollController,
+            );
+          }
+        },
+      ),
     );
-  }
-}
-
-class NewsList extends StatelessWidget {
-  Future<RssFeed> getFeed() async {
-    var client = new http.Client();
-    // RSS feed
-    var response =
-        await client.get('https://www.nasa.gov/rss/dyn/breaking_news.rss');
-    var channel = new RssFeed.parse(response.body);
-    return channel;
   }
 
   @override
-  Widget build(BuildContext context) {
-    return FutureBuilder(
-      future: getFeed(),
-      builder: (context, feed) {
-        if (feed.connectionState == ConnectionState.waiting)
-          return NativeLoadingIndicator(
-            center: true,
-            text: Text('Loading...'),
-          );
+  void dispose() {
+    _postBloc.dispose();
+    super.dispose();
+  }
 
-        if (feed.connectionState == ConnectionState.done) {
-          if (feed.data?.items == null || feed.data.items.isEmpty)
-            return Center(child: Text('No Feed Found'));
-
-          return ListView.builder(
-            itemCount: feed.data.items.length ?? 0,
-            itemBuilder: (BuildContext context, int index) {
-              return ListTile(
-                title: Text(feed.data.items[index].title),
-                subtitle: Text(feed.data.items[index].description),
-              );
-            },
-          );
-        }
-      },
-    );
+  void _onScroll() {
+    final maxScroll = _scrollController.position.maxScrollExtent;
+    final currentScroll = _scrollController.position.pixels;
+    if (maxScroll - currentScroll <= _scrollThreshold) {
+      _postBloc.dispatch(Fetch());
+    }
   }
 }
