--- lib/ui/post/screen.dart
@@ -1,81 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:http/http.dart' as http;
-
-import '../../bloc/bloc.dart';
-import '../app/bottom_loader.dart';
-import 'item.dart';
-
-class ArticlesScreen extends StatefulWidget {
-  @override
-  _ArticlesScreenState createState() => _ArticlesScreenState();
-}
-
-class _ArticlesScreenState extends State<ArticlesScreen> {
-  final _scrollController = ScrollController();
-  final PostBloc _postBloc = PostBloc(httpClient: http.Client());
-  final _scrollThreshold = 200.0;
-
-  _ArticlesScreenState() {
-    _scrollController.addListener(_onScroll);
-    _postBloc.dispatch(Fetch());
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text("Space News"),
-      ),
-      body: BlocBuilder(
-        bloc: _postBloc,
-        builder: (BuildContext context, PostState state) {
-          if (state is PostUninitialized) {
-            return Center(
-              child: CircularProgressIndicator(),
-            );
-          }
-          if (state is PostError) {
-            return Center(
-              child: Text('failed to fetch posts'),
-            );
-          }
-          if (state is PostLoaded) {
-            if (state.posts.isEmpty) {
-              return Center(
-                child: Text('no posts'),
-              );
-            }
-            return ListView.builder(
-              itemBuilder: (BuildContext context, int index) {
-                if (index >= state.posts.length) {
-                  return BottomLoader();
-                } else {
-                  return PostWidget(post: state.posts[index]);
-                }
-              },
-              itemCount: state.hasReachedMax
-                  ? state.posts.length
-                  : state.posts.length + 1,
-              controller: _scrollController,
-            );
-          }
-        },
-      ),
-    );
-  }
-
-  @override
-  void dispose() {
-    _postBloc.dispose();
-    super.dispose();
-  }
-
-  void _onScroll() {
-    final maxScroll = _scrollController.position.maxScrollExtent;
-    final currentScroll = _scrollController.position.pixels;
-    if (maxScroll - currentScroll <= _scrollThreshold) {
-      _postBloc.dispatch(Fetch());
-    }
-  }
-}
